further project package structuring
