+++ src/swurg/model/RESTful.java
@@ -14,7 +14,7 @@
 #    limitations under the License. 
 */
 
-package swurg;
+package swurg.model;
 
 import com.google.gson.JsonObject;
